Fixes issue #67 (backstack)

+++ MediaBrowser.WindowsPhone8/Views/FindServerView.xaml.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.Phone.Controls;
+﻿using System.Windows.Navigation;
 
 namespace MediaBrowser.WindowsPhone.Views
 {
@@ -16,8 +16,10 @@ namespace MediaBrowser.WindowsPhone.Views
         protected override void OnNavigatedFrom(System.Windows.Navigation.NavigationEventArgs e)
         {
             base.OnNavigatedTo(e);
-            if (NavigationService.CanGoBack)
+            if (NavigationService.CanGoBack && e.NavigationMode != NavigationMode.Back)
+            {
                 NavigationService.RemoveBackEntry();
+            }
         }
     }
 }
